refactor: input state들 scenarioInputDataAtom으로 변경
구조 변경
