+++ src/components/CoinScenarioForm/DateInput.jsx
@@ -1,19 +1,21 @@
 import React from 'react';
 import DatePicker from 'react-datepicker';
 import 'react-datepicker/dist/react-datepicker.css';
-import PropTypes from 'prop-types';
 import useResponsiveView from 'hooks/useResponsiveView';
-
+import { useRecoilState } from 'recoil';
 /** @jsxImportSource @emotion/react */
 import whiteInvertedTriangleIcon from 'assets/white-inverted-triangle.svg';
 import invertedTriangleIcon from 'assets/inverted-triangle.svg';
 import ko from 'date-fns/locale/ko';
+import { selectedDateAtom } from 'recoils/scenarioInputData/scenarioInputDataAtom';
 import { coinScenarioInputStyle } from './coinScenarioInputStyle';
 
-const DateInput = ({ selectedDate, onSelectedDate }) => {
+const DateInput = () => {
+  const [selectedDate, setSelectedDate] = useRecoilState(selectedDateAtom);
+
   const viewportType = useResponsiveView();
   const handleDateChange = (date) => {
-    onSelectedDate(date);
+    setSelectedDate(date);
   };
 
   return (
@@ -33,9 +35,5 @@ const DateInput = ({ selectedDate, onSelectedDate }) => {
       </p>
     </div>
   );
-};
-DateInput.propTypes = {
-  onSelectedDate: PropTypes.func.isRequired,
-  selectedDate: PropTypes.instanceOf(Date),
 };
 export default DateInput;
