+++ src/components/CoinScenarioForm/ScenarioDescription.jsx
@@ -6,6 +6,7 @@ import useFormattedPrice from 'hooks/useFormattedPrice';
 import { BASE_CURRENCY, EXCHANGE_RATE } from 'utils/constants';
 import localeCurrencyAtom from 'recoils/localeCurrency/localeCurrencyAtom';
 import { useRecoilValue } from 'recoil';
+import { selectedDateAtom, buyPriceAtom, selectedCoinAtom } from 'recoils/scenarioInputData/scenarioInputDataAtom';
 
 const containerStyle = css`
   color: var(--gray5);
@@ -45,12 +46,20 @@ const strongText = css`
 `;
 
 const ScenarioDescription = ({
-  year, month, day, selectedCoin, price, onBottomSheetClick,
+  onBottomSheetClick,
 }) => {
+  const price = useRecoilValue(buyPriceAtom);
+  const selectedDate = useRecoilValue(selectedDateAtom);
+  const selectedCoin = useRecoilValue(selectedCoinAtom);
   const formatPrice = useFormattedPrice();
   const localeCurrency = useRecoilValue(localeCurrencyAtom);
   const convertedPrice = localeCurrency === BASE_CURRENCY ? price : price / EXCHANGE_RATE[`${BASE_CURRENCY}TO${localeCurrency}`];
   const formattedPrice = formatPrice(convertedPrice);
+
+  const [year, month, day] = selectedDate
+    ? [selectedDate.getFullYear().toString(),
+      (selectedDate.getMonth() + 1).toString(), selectedDate.getDate().toString()]
+    : ['0000', '00', '00'];
 
   return (
     <h1 css={headingStyle}>
@@ -79,17 +88,6 @@ const ScenarioDescription = ({
 };
 
 ScenarioDescription.propTypes = {
-  year: PropTypes.string.isRequired,
-  month: PropTypes.string.isRequired,
-  day: PropTypes.string.isRequired,
-
-  selectedCoin: PropTypes.shape({
-    id: PropTypes.string.isRequired,
-    name: PropTypes.string.isRequired,
-    image: PropTypes.string.isRequired,
-  }).isRequired,
-
-  price: PropTypes.number.isRequired,
   onBottomSheetClick: PropTypes.func,
 };
